Break out Troubleshoot row for storage panel
